refactor: change cjs to esm, remove unnecessary backticks and commented-out code and add more typings
