+++ config/db.ts
@@ -1,0 +1,23 @@
+import mongoose from "mongoose";
+
+const options = {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+  serverSelectionTimeoutMS: 5000,
+};
+const uri =
+  process.env.NODE_ENV === "production"
+    ? process.env.MONGODB_URI
+    : process.env.MONGODB_URI_LOCAL;
+
+const connectDB = async (): Promise<void> => {
+  try {
+    const conn = await mongoose.connect(uri || "", options);
+    console.log(`MongoDB Connected: ${conn.connection.host}`);
+  } catch (error) {
+    console.log(error);
+    process.exit(1);
+  }
+};
+
+export default connectDB;
