+++ routes/allUploadsRoute.ts
@@ -1,11 +1,8 @@
 import express from "express";
 const router = express.Router();
-const {
-  getAllDocs,
-  deleteDoc,
-} = require("../controllers/allUploadsController");
+import { getAllDocs, deleteDoc } from "../controllers/allUploadsController";
 
-router.get(`/`, getAllDocs);
+router.get("/", getAllDocs);
 router.delete("/:doc/:id", deleteDoc);
 
-module.exports = router;
+export default router;
