+++ routes/userRoute.ts
@@ -7,4 +7,4 @@ router.post("/", registerUser);
 router.post("/login", loginUser);
 router.get("/me", protect, getMe);
 
-module.exports = router;
+export default router;
